LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/inserted-or-removed-expected.txt



PASS The last selected OPTION should win; Inserted by parser
FAIL The last selected OPTION should win; Inserted by DOM API assert_equals: expected "Second" but got "First"
PASS The last selected OPTION should win; Inserted by innerHTML
